refactor: split auth.ts server actions by responsibility (#972)
move escrow status actions to app/actions/escrow/escrow-status-actions.ts
and password reset actions to app/actions/auth/password-reset-actions.ts,
keeping session actions in auth.ts. update call sites accordingly. no
behavior changes.

Closes #955
