+++ apps/web/app/actions/auth.ts
@@ -1,57 +1,25 @@
 'use server'
 
 import { createSupabaseServerClient } from '@packages/lib/supabase-server'
-import type { Database } from '@services/supabase'
 import type { AuthError } from '@supabase/supabase-js'
-import { revalidatePath } from 'next/cache'
 import { cookies } from 'next/headers'
 import { redirect } from 'next/navigation'
-import { validateCsrfToken } from '~/app/actions/csrf'
 import { AuthErrorHandler } from '~/lib/auth/error-handler'
 import {
 	enforceRateLimit,
-	getAuthenticatedSession,
-	requireAdminSession,
 	requireAuthenticatedSession,
-	type ServerActionFailure,
 	toServerActionFailure,
 	validateInput,
 } from '~/lib/auth/server-action-auth'
 import { Logger } from '~/lib/logger'
 import {
 	createSessionInputSchema,
-	requestResetAccountInputSchema,
-	resetPasswordInputSchema,
 	updateDeviceWithDeployeeInputSchema,
-	updateEscrowFinancialsInputSchema,
-	updateEscrowMilestoneInputSchema,
-	updateEscrowStatusInputSchema,
 } from '~/lib/schemas/server-actions.schemas'
 import type { AuthResponse } from '~/lib/types/auth'
 
-type Tables = Database['public']['Tables']
-type EscrowRecord = Tables['escrow_status']['Row']
-
-type EscrowResponse = {
-	success: boolean
-	message: string
-	data?: EscrowRecord | EscrowRecord[] | null
-	error?: string | null
-}
-
 const logger = new Logger()
 const errorHandler = new AuthErrorHandler(logger)
-
-function escrowFailureFromAction(
-	failure: ServerActionFailure,
-	fallbackMessage: string,
-): EscrowResponse {
-	return {
-		success: false,
-		message: failure.error || fallbackMessage,
-		error: failure.error,
-	}
-}
 
 export async function createSessionAction(input: {
 	userId: string
@@ -176,98 +144,6 @@ export async function signOutAction(): Promise<void> {
 	}
 }
 
-export async function requestResetAccountAction(formData: FormData): Promise<void> {
-	if (!(await validateCsrfToken(formData.get('csrfToken')?.toString()))) {
-		redirect('/reset-account?error=Invalid CSRF token')
-	}
-
-	let email: string
-	try {
-		const validated = validateInput(
-			requestResetAccountInputSchema,
-			{ email: formData.get('email')?.toString() },
-			'requestResetAccountAction',
-		)
-		email = validated.email
-	} catch (_error) {
-		redirect('/reset-account?error=A valid email is required')
-	}
-
-	try {
-		await enforceRateLimit(email.toLowerCase(), 'request_reset_account')
-	} catch (_error) {
-		redirect('/reset-account?error=Too many requests. Try again later.')
-	}
-
-	logger.warn({
-		eventType: 'RESET_ACCOUNT_NOT_IMPLEMENTED',
-		email,
-	})
-	redirect(
-		'/reset-account?error=Account recovery is temporarily unavailable. Please contact support.',
-	)
-}
-
-export async function resetPasswordAction(formData: FormData): Promise<void> {
-	if (!(await validateCsrfToken(formData.get('csrfToken')?.toString()))) {
-		redirect('/reset-password?error=Invalid CSRF token')
-	}
-
-	const session = await getAuthenticatedSession()
-	if (!session) {
-		redirect(
-			'/reset-password?error=You must follow the recovery link sent to your email before resetting your password.',
-		)
-	}
-
-	let password: string
-	try {
-		const validated = validateInput(
-			resetPasswordInputSchema,
-			{
-				password: formData.get('password')?.toString() ?? '',
-				confirmPassword: formData.get('confirmPassword')?.toString() ?? '',
-			},
-			'resetPasswordAction',
-		)
-		password = validated.password
-	} catch (_error) {
-		redirect(
-			'/reset-password?error=Password must be at least 8 characters and match the confirmation.',
-		)
-	}
-
-	try {
-		await enforceRateLimit(session.user.id, 'reset_password')
-	} catch (_error) {
-		redirect('/reset-password?error=Too many requests. Try again later.')
-	}
-
-	const supabase = await createSupabaseServerClient()
-
-	try {
-		const { error } = await supabase.auth.updateUser({ password })
-
-		if (error) {
-			const response = errorHandler.handleAuthError(error, 'reset_password')
-			redirect(`/reset-password?error=${encodeURIComponent(response.message)}`)
-		}
-
-		redirect('/sign-in?success=Password updated successfully')
-	} catch (error) {
-		if (
-			error &&
-			typeof error === 'object' &&
-			'digest' in error &&
-			(error.digest as string | undefined)?.startsWith('NEXT_REDIRECT')
-		) {
-			throw error
-		}
-		const response = errorHandler.handleAuthError(error as AuthError, 'reset_password')
-		redirect(`/reset-password?error=${encodeURIComponent(response.message)}`)
-	}
-}
-
 export async function checkAuthStatus(): Promise<AuthResponse> {
 	const supabase = await createSupabaseServerClient()
 
@@ -295,282 +171,6 @@ export async function checkAuthStatus(): Promise<AuthResponse> {
 		}
 	} catch (error) {
 		return errorHandler.handleAuthError(error as AuthError, 'check_auth')
-	}
-}
-
-export async function updateEscrowStatusAction(
-	id: string,
-	newStatus: string,
-): Promise<EscrowResponse> {
-	try {
-		await requireAdminSession('updateEscrowStatusAction')
-	} catch (error) {
-		return escrowFailureFromAction(toServerActionFailure(error, 'Forbidden'), 'Forbidden')
-	}
-
-	let validated: { id: string; newStatus: EscrowRecord['status'] }
-	try {
-		validated = validateInput(
-			updateEscrowStatusInputSchema,
-			{ id, newStatus },
-			'updateEscrowStatusAction',
-		) as { id: string; newStatus: EscrowRecord['status'] }
-	} catch (error) {
-		return escrowFailureFromAction(toServerActionFailure(error, 'Invalid input'), 'Invalid input')
-	}
-
-	const supabase = await createSupabaseServerClient()
-
-	try {
-		const { data, error } = await supabase
-			.from('escrow_status')
-			.update({
-				status: validated.newStatus,
-				last_updated: new Date().toISOString(),
-			})
-			.eq('id', validated.id)
-			.select()
-			.single()
-
-		if (error) throw error
-
-		revalidatePath('/admin/escrow')
-		return {
-			success: true,
-			message: `Status updated to ${validated.newStatus}`,
-			data,
-		}
-	} catch (error) {
-		logger.error({
-			eventType: 'ESCROW_STATUS_UPDATE_ERROR',
-			error: error instanceof Error ? error.message : 'Unknown error',
-			id: validated.id,
-			newStatus: validated.newStatus,
-		})
-		return {
-			success: false,
-			message: 'Failed to update escrow status',
-			error: error instanceof Error ? error.message : 'Unknown error',
-		}
-	}
-}
-
-export async function updateEscrowMilestoneAction(
-	id: string,
-	current: number,
-	completed: number,
-): Promise<EscrowResponse> {
-	try {
-		await requireAdminSession('updateEscrowMilestoneAction')
-	} catch (error) {
-		return escrowFailureFromAction(toServerActionFailure(error, 'Forbidden'), 'Forbidden')
-	}
-
-	let validated: { id: string; current: number; completed: number }
-	try {
-		validated = validateInput(
-			updateEscrowMilestoneInputSchema,
-			{ id, current, completed },
-			'updateEscrowMilestoneAction',
-		)
-	} catch (error) {
-		return escrowFailureFromAction(toServerActionFailure(error, 'Invalid input'), 'Invalid input')
-	}
-
-	const supabase = await createSupabaseServerClient()
-
-	try {
-		const { data, error } = await supabase
-			.from('escrow_status')
-			.update({
-				current_milestone: validated.current,
-				metadata: {
-					milestoneStatus: {
-						current: validated.current,
-						completed: validated.completed,
-					},
-				},
-				last_updated: new Date().toISOString(),
-			})
-			.eq('id', validated.id)
-			.select()
-			.single()
-
-		if (error) throw error
-
-		revalidatePath('/admin/escrow')
-		return {
-			success: true,
-			message: 'Milestone updated successfully',
-			data,
-		}
-	} catch (error) {
-		logger.error({
-			eventType: 'ESCROW_MILESTONE_UPDATE_ERROR',
-			error: error instanceof Error ? error.message : 'Unknown error',
-			id: validated.id,
-			current: validated.current,
-			completed: validated.completed,
-		})
-		return {
-			success: false,
-			message: 'Failed to update milestone',
-			error: error instanceof Error ? error.message : 'Unknown error',
-		}
-	}
-}
-
-export async function updateEscrowFinancialsAction(
-	id: string,
-	funded: number,
-	released: number,
-): Promise<EscrowResponse> {
-	try {
-		await requireAdminSession('updateEscrowFinancialsAction')
-	} catch (error) {
-		return escrowFailureFromAction(toServerActionFailure(error, 'Forbidden'), 'Forbidden')
-	}
-
-	let validated: { id: string; funded: number; released: number }
-	try {
-		validated = validateInput(
-			updateEscrowFinancialsInputSchema,
-			{ id, funded, released },
-			'updateEscrowFinancialsAction',
-		)
-	} catch (error) {
-		return escrowFailureFromAction(toServerActionFailure(error, 'Invalid input'), 'Invalid input')
-	}
-
-	const supabase = await createSupabaseServerClient()
-
-	try {
-		const { data, error } = await supabase
-			.from('escrow_status')
-			.update({
-				total_funded: validated.funded,
-				total_released: validated.released,
-				last_updated: new Date().toISOString(),
-			})
-			.eq('id', validated.id)
-			.select()
-			.single()
-
-		if (error) throw error
-
-		revalidatePath('/admin/escrow')
-		return {
-			success: true,
-			message: 'Financials updated successfully',
-			data,
-		}
-	} catch (error) {
-		logger.error({
-			eventType: 'ESCROW_FINANCIALS_UPDATE_ERROR',
-			error: error instanceof Error ? error.message : 'Unknown error',
-			id: validated.id,
-			funded: validated.funded,
-			released: validated.released,
-		})
-		return {
-			success: false,
-			message: 'Failed to update financials',
-			error: error instanceof Error ? error.message : 'Unknown error',
-		}
-	}
-}
-
-export async function getEscrowRecordsAction(): Promise<EscrowResponse> {
-	try {
-		await requireAdminSession('getEscrowRecordsAction')
-	} catch (error) {
-		return escrowFailureFromAction(toServerActionFailure(error, 'Forbidden'), 'Forbidden')
-	}
-
-	const supabase = await createSupabaseServerClient()
-
-	try {
-		const { data, error } = await supabase
-			.from('escrow_status')
-			.select('*')
-			.order('last_updated', { ascending: false })
-
-		if (error) throw error
-
-		return {
-			success: true,
-			message: 'Records fetched successfully',
-			data,
-		}
-	} catch (error) {
-		logger.error({
-			eventType: 'ESCROW_RECORDS_FETCH_ERROR',
-			error: error instanceof Error ? error.message : 'Unknown error',
-		})
-		return {
-			success: false,
-			message: 'Failed to fetch records',
-			error: error instanceof Error ? error.message : 'Unknown error',
-		}
-	}
-}
-
-export async function insertTestEscrowRecordAction(): Promise<EscrowResponse> {
-	if (process.env.NODE_ENV === 'production') {
-		return {
-			success: false,
-			message: 'Test record insertion is disabled in production',
-			error: 'Forbidden',
-		}
-	}
-
-	try {
-		await requireAdminSession('insertTestEscrowRecordAction')
-	} catch (error) {
-		return escrowFailureFromAction(toServerActionFailure(error, 'Forbidden'), 'Forbidden')
-	}
-
-	const supabase = await createSupabaseServerClient()
-
-	try {
-		const { data, error } = await supabase
-			.from('escrow_status')
-			.insert([
-				{
-					escrow_id: `test-${Date.now()}`,
-					status: 'NEW',
-					current_milestone: 1,
-					total_funded: 1000,
-					total_released: 0,
-					metadata: {
-						milestoneStatus: {
-							total: 3,
-							completed: 0,
-						},
-					},
-				},
-			])
-			.select()
-			.single()
-
-		if (error) throw error
-
-		revalidatePath('/admin/escrow')
-		return {
-			success: true,
-			message: 'Test record inserted successfully',
-			data,
-		}
-	} catch (error) {
-		logger.error({
-			eventType: 'ESCROW_TEST_RECORD_INSERT_ERROR',
-			error: error instanceof Error ? error.message : 'Unknown error',
-		})
-		return {
-			success: false,
-			message: 'Failed to insert test record',
-			error: error instanceof Error ? error.message : 'Unknown error',
-		}
 	}
 }
